STERICMS-876: consolidate styles, enhance layout, and implement cookie functionality for dismissal
