+++ blocks/banner/banner.css
@@ -28,6 +28,14 @@
   z-index: 100;
   width: 100%;
   bottom: 0;
+  background-color: var(--color-light-gray);
+  min-height: 64px;
+  max-width: none;
+  padding: 0 1rem;
+}
+
+.block.banner.fixed-to-bottom p {
+  color: var(--color-text-primary);
 }
 
 .banner-container:has(.fixed-to-top) {
@@ -38,7 +46,18 @@
   z-index: 100;
 }
 
-.block.banner > .banner-content {
+.block.banner > .banner-content-wrapper {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  max-width: var(--section-width);
+  margin: 0 auto;
+  width: 100%;
+  justify-content: space-between;
+  padding: 1rem;
+}
+
+.block.banner > .banner-content-wrapper > .banner-content {
   position: sticky;
   display: flex;
   flex-direction: row;
@@ -49,76 +68,48 @@
   align-items: center;
 }
 
-@media screen and (min-width: 576px /* breakpoint-s */) {
-  .block.banner > .banner-content {
-    gap: 4rem;
-  }
+.block.banner.fixed-to-bottom > .banner-content-wrapper > .banner-content {
+  justify-content: flex-start;
+  margin: 0 auto;
 }
 
-@media screen and (min-width: 992px /* breakpoint-l */) {
-  .block.banner > .banner-content {
-    gap: 10rem;
-  }
-
-  .block.banner > span {
-    display: none; /* Hide the span element */
-    width: 50px;
-    gap: 50px;
-    color: var(--color-white);
-    margin-left: 10px;
-  }
-}
-
-.block.banner > .banner-content > .banner-content-page * {
-  font-size: 16px;
-  font-weight: 700;
-  line-height: 16px;
-  letter-spacing: 0.32px;
-  text-decoration: underline;
-}
-
-.block.banner > .banner-content > .banner-content-page {
-  display: none;
-}
-
-.block.banner > .banner-content > .banner-content-page.active {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.block.banner > .banner-content > .banner-content-page .icon {
-  --icon-color: var(--color-white);
-
-  width: 12px;
-  height: 12px;
-}
-
-.block.banner > .banner-actions {
+.block.banner > .banner-content-wrapper > .banner-actions {
   display: flex;
   flex-direction: column;
   justify-content: center;
+  align-items: center;
+  margin-left: 1rem;
 }
 
-.block.banner > .banner-actions > .close-button {
+.block.banner > .banner-content-wrapper > .banner-actions > .close-button {
   margin: 0;
   padding: 0;
   width: 32px;
   height: 32px;
   display: flex;
   justify-content: center;
+  align-items: center;
   flex-direction: column;
   color: unset;
   background-color: unset;
   border: 1px solid var(--color-white);
   border-radius: var(--border-radius-l);
   font-size: 12px;
+  font-weight: 400;
 }
 
-.block.banner > .banner-actions > .close-button::before {
+.block.banner.fixed-to-bottom > .banner-content-wrapper > .banner-actions > .close-button {
+  border-color: var(--color-text-primary);
+}
+
+.block.banner > .banner-content-wrapper > .banner-actions > .close-button::before {
   content: "X";
   color: var(--color-white);
   margin: auto;
+}
+
+.block.banner.fixed-to-bottom > .banner-content-wrapper > .banner-actions > .close-button::before {
+  color: var(--color-text-primary);
 }
 
 .block.banner p:has(a) a:any-link,
@@ -134,14 +125,14 @@
 }
 
 .fa-chevron-left::before {
-  content: "";
+  content: "";
   font-family: "Font Awesome\ 6 Free", "Font Awesome\ 5 Free", serif;
   margin-left: 8px;
   color: var(--color-white);
 }
 
 .fa-chevron-right::after {
-  content: "";
+  content: "";
   font-family: "Font Awesome\ 6 Free", "Font Awesome\ 5 Free", serif;
   margin-right: 8px;
   color: var(--color-white);
@@ -168,4 +159,49 @@ button.fa-chevron-right {
   overflow: unset;
   text-overflow: unset;
   border-radius: unset;
+}
+
+@media screen and (min-width: 576px /* breakpoint-s */) {
+  .block.banner > .banner-content-wrapper > .banner-content {
+    gap: 4rem;
+  }
+}
+
+@media screen and (min-width: 992px /* breakpoint-l */) {
+  .block.banner > .banner-content-wrapper > .banner-content {
+    gap: 10rem;
+  }
+
+  .block.banner > span {
+    display: none; /* Hide the span element */
+    width: 50px;
+    gap: 50px;
+    color: var(--color-white);
+    margin-left: 10px;
+  }
+}
+
+.block.banner > .banner-content-wrapper > .banner-content > .banner-content-page * {
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 16px;
+  letter-spacing: 0.32px;
+  text-decoration: underline;
+}
+
+.block.banner > .banner-content-wrapper > .banner-content > .banner-content-page {
+  display: none;
+}
+
+.block.banner > .banner-content-wrapper > .banner-content > .banner-content-page.active {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.block.banner > .banner-content-wrapper > .banner-content > .banner-content-page .icon {
+  --icon-color: var(--color-white);
+
+  width: 12px;
+  height: 12px;
 }
