--- blocks/privacy-banner/privacy-banner.css
@@ -1,90 +1,0 @@
-.block.privacy-banner {
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  z-index: 100;
-  background: var(--color-light-gray);
-  color: var(--color-text-primary);
-  font-family: inherit;
-  box-shadow: 0 -2px 4px rgba(0,0,0,0.05);
-  padding: 0;
-}
-
-.privacy-banner-columns {
-  display: flex;
-  justify-content: space-between;
-  align-items: stretch;
-  width: 100%;
-  max-width: var(--section-width);
-  margin: 0 auto;
-  padding: 1rem 2rem;
-  gap: 4rem;
-}
-
-.privacy-banner-left {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.privacy-title {
-  font-weight: bold;
-  font-size: 1.1rem;
-  margin-bottom: 0.25rem;
-}
-
-.privacy-date {
-  font-size: 1rem;
-  color: var(--color-text-primary);
-}
-
-.privacy-banner-right {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  flex: 1;
-}
-
-.privacy-link {
-  font-weight: bold;
-  color: #1a5a37;
-  text-decoration: underline;
-  font-size: 1.1rem;
-  flex: 1;
-  height: 100%;
-}
-
-.close-button {
-  margin: 0;
-  padding: 0;
-  width: 32px;
-  height: 32px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  color: var(--color-text-primary);
-  background-color: unset;
-  border: 1px solid var(--color-text-primary);
-  border-radius: var(--border-radius-l);
-  font-size: 12px;
-}
-
-.close-button::before {
-  content: "X";
-  color: var(--color-text-primary);
-  margin: auto;
-}
-
-@media (max-width: 700px) {
-  .privacy-banner-columns {
-    flex-direction: column-reverse;
-    padding: 0.75rem 1rem;
-    gap: 0.5rem;
-  }
-  .privacy-banner-right {
-    justify-content: flex-end;
-    margin-top: 0.5rem;
-  }
-}
